# api/openapi.yaml
openapi: 3.1.0
info:
  title: kmdn API
  version: 0.1.0
  description: |
    REST API of the kmdn server. Paths are relative to `/api/v1`.

    Browser clients authenticate with the `kmdn_session` cookie. Unsafe
    methods (POST, PUT, PATCH, DELETE) must send the `X-Kmdn-CSRF` header with
    the value of the `kmdn_csrf` cookie (also returned as `csrf_token` by
    `GET /me`). Errors use RFC 9457 problem details with a stable `code`.
  license:
    name: AGPL-3.0
servers:
  - url: /api/v1
...
paths:
...
  /orgs:
    get:
      tags: [orgs]
      operationId: listOrgs
      summary: The organizations the caller belongs to, with their role
      responses:
        "200":
          description: Organizations
          content:
            application/json:
              schema:
                type: object
                required: [items, mode, can_create, org_forges]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/OrgView" } }
                  mode: { type: string, enum: [single, multi] }
                  can_create: { type: boolean, description: "The caller may create an organization" }
                  org_forges: { type: boolean, description: "Org admins can add their own forges (policy)" }
    post:
      tags: [orgs]
      operationId: createOrg
      summary: Create an organization (multi mode; who may is orgs.allow_create); the caller becomes its owner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrgInput" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/OrgView" } } } }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /orgs/{org}:
    get:
      tags: [orgs]
      operationId: getOrg
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      responses:
        "200": { description: Organization, content: { application/json: { schema: { $ref: "#/components/schemas/OrgView" } } } }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [orgs]
      operationId: updateOrg
      summary: Rename the org (admins) or change its address (owners)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
      responses:
        "200": { description: Organization, content: { application/json: { schema: { $ref: "#/components/schemas/OrgView" } } } }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /orgs/{org}/members:
    get:
      tags: [orgs]
      operationId: listOrgMembers
      summary: The org's members with their roles (org admins)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/OrgMember" } } } }
  /orgs/{org}/members/{user}:
    patch:
      tags: [orgs]
      operationId: updateOrgMember
      summary: Change a member's role or deactivate them in this org (admins; owners for owners)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
        - { name: user, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [member, admin, owner] }
                status: { type: string, enum: [active, deactivated] }
      responses:
        "200": { description: Changed, content: { application/json: { schema: { type: object } } } }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [orgs]
      operationId: removeOrgMember
      summary: Remove a member (admins), or leave the org (yourself)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
        - { name: user, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Removed, content: { application/json: { schema: { type: object } } } }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /orgs/{org}/admin/settings:
    get:
      tags: [orgs]
      operationId: getOrgSettings
      summary: The org's settings, and those the deployment manages (read-only)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      responses:
        "200": { description: Settings, content: { application/json: { schema: { $ref: "#/components/schemas/OrgSettingsView" } } } }
    patch:
      tags: [orgs]
      operationId: updateOrgSettings
      summary: Change org settings (org admins); managed fields are refused
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrgSettings" }
      responses:
        "200": { description: Settings, content: { application/json: { schema: { $ref: "#/components/schemas/OrgSettingsView" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /orgs/{org}/admin/domains:
    get:
      tags: [orgs]
      operationId: listOrgDomains
      summary: Email domains the org claims (org admins)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      responses:
        "200":
          description: Domains
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/OrgDomain" } } } }
    post:
      tags: [orgs]
      operationId: addOrgDomain
      summary: Claim an email domain; people there join the org on their own once it's verified
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [domain], properties: { domain: { type: string } } }
      responses:
        "201": { description: Claimed, content: { application/json: { schema: { $ref: "#/components/schemas/OrgDomain" } } } }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /orgs/{org}/admin/domains/{domain}:
    delete:
      tags: [orgs]
      operationId: removeOrgDomain
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
        - { name: domain, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
        "404": { $ref: "#/components/responses/Problem" }
...
  /admin/orgs/{org}:
    patch:
      tags: [admin]
      operationId: updateOrgStatus
      summary: Suspend or resume an organization (instance admins)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [status], properties: { status: { type: string, enum: [active, suspended] }, reason: { type: string, description: "Shown to members while suspended" } } }
      responses:
        "200": { description: Organization, content: { application/json: { schema: { $ref: "#/components/schemas/Org" } } } }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/orgs/{org}/domains/{domain}/verify:
    post:
      tags: [admin]
      operationId: verifyOrgDomain
      summary: Mark an org's domain verified (instance admins)
      parameters:
        - { $ref: "#/components/parameters/OrgSlug" }
        - { name: domain, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Verified }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
    OrgSlug:
      name: org
      in: path
      required: true
      description: The organization's slug
      schema: { type: string }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: kmdn_session
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Org:
      type: object
      required: [id, slug, name, status, created_at]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        status: { type: string, enum: [active, suspended, deleting] }
        status_reason: { type: string, description: "Why a suspended org is read-only" }
        created_at: { type: string, format: date-time }
    OrgView:
      allOf:
        - { $ref: "#/components/schemas/Org" }
        - type: object
          required: [role]
          properties:
            role: { type: string, enum: [member, admin, owner, ""], description: "Empty for an instance admin outside the org" }
            sign_in:
              type: object
              description: Set when this session can't act in the org until the person signs in the way the org asks
              required: [org_id, message]
              properties:
                org_id: { type: string }
                message: { type: string }
                url: { type: string, description: Starts a sign-in the org accepts }
    OrgInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        slug: { type: string, description: "Made from the name when empty" }
    OrgSettings:
      type: object
      properties:
        assistant: { type: boolean, description: "AI features for the org's repositories" }
        monthly_tokens: { type: integer, description: "The org's AI budget per calendar month (UTC); 0: none of its own" }
    OrgSettingsView:
      type: object
      required: [settings, locked]
      properties:
        settings: { $ref: "#/components/schemas/OrgSettings" }
        locked: { type: array, items: { type: string }, description: "Fields managed by the deployment" }
    OrgDomain:
      type: object
      required: [domain, org_id, created_at]
      properties:
        domain: { type: string }
        org_id: { type: string }
        verified_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    OrgMember:
      type: object
      required: [id, name, email, role, status, joined_at]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [member, admin, owner] }
        status: { type: string, enum: [active, deactivated] }
        joined_at: { type: integer, description: Unix milliseconds }
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        params: { type: object, additionalProperties: true }
...
security:
  - session: []